+++ tests/test_catalog.py
@@ -1,24 +1,16 @@
 """
 Tests for the catalog methods.
 """
-from unittest.mock import patch
-import json
-import sys
-import os
-import pytest
+
 from cortexapps_cli.cortex import cli
-import yaml
+import json
 import pytest
 
-def test_create_service(capsys):
-    cli(["catalog", "create", "-f", "tests/test_catalog_service.yaml"])
+def test_catalog_create_service(capsys):
+    cli(["catalog", "create", "-f", "tests/test_catalog_create_service.yaml"])
 
 def test_retrieve_service(capsys):
-    cli(["catalog", "descriptor", "-y", "-t", "test-service"])
-    out, err = capsys.readouterr()
-    with open('tests/test_catalog_service.yaml', 'r') as f:
-       y = f.read()
-    assert out == y, "Contents of source yaml and retrieved descriptor should be identical"
+    cli(["catalog", "descriptor", "-y", "-t", "cli-test-service"])
 
 def test_dryrun(capsys):
     with pytest.raises(SystemExit) as excinfo:
@@ -27,7 +19,7 @@ def test_dryrun(capsys):
         assert json.loads(out)['type'] == "BAD_REQUEST"
 
 def test_details(capsys):
-    cli(["catalog", "details", "-t", "test-service"])
+    cli(["catalog", "details", "-t", "cli-test-service"])
 
 def test_list(capsys):
     cli(["catalog", "list"])
@@ -36,18 +28,17 @@ def test_list_with_parms(capsys):
     cli(["catalog", "list", "-g", "corona-spokesperson", "-d", "1", "-t", "service", "-a", "-m" ])
     out, err = capsys.readouterr()
     out = json.loads(out)
-    assert str(out['entities'][0]['tag']) == "test-service", "x-cortex-tag should be test-service"
+    assert any(service['tag'] == 'cli-test-service-with-groups' for service in out['entities'])
     assert not(out['entities'][0]['metadata'][0]["key"] is None), "Custom metadata should have been in result"
 
+# Archiving a service can impact it from being seen by other operations.  Should probably be done with a separate
+# service
 def test_archive():
-    cli(["catalog", "archive", "-t", "test-service"])
-
-def test_unarchive():
-    cli(["catalog", "unarchive", "-t", "test-service"])
+    cli(["catalog", "archive", "-t", "cli-test-service"])
+    cli(["catalog", "unarchive", "-t", "cli-test-service"])
 
 def test_list_by_team(capsys):
     cli(["catalog", "list", "-o", "test-team-1" ])
     out, err = capsys.readouterr()
     out = json.loads(out)
-    assert str(out['entities'][0]['tag']) == "test-service"
-
+    assert any(service['tag'] == 'cli-test-service-with-groups' for service in out['entities'])