+++ tests/test_integrations_github.py
@@ -42,17 +42,7 @@ def github_app_input(tmp_path):
     f.write_text(content)
     return f
 
-def test_integrations_github_add(capsys, tmp_path):
-    cli(["integrations", "github", "get", "-a", "github-test-3"])
-    out, err = capsys.readouterr()
-    out = json.loads(out)
-    if (out['type'] != "NOT_FOUND"):
-        cli(["integrations", "github", "delete", "-a", "github-test-3"])
-
-    f = github_app_input(tmp_path)
-    cli(["integrations", "github", "add", "-f", str(f)])
-
-def test_integrations_github_add_personal(capsys, tmp_path):
+def test_integrations_github_personal(capsys, tmp_path):
     cli(["integrations", "github", "get-personal", "-a", "github-personal-test-001"])
     out, err = capsys.readouterr()
     out = json.loads(out)
@@ -62,29 +52,31 @@ def test_integrations_github_add_personal(capsys, tmp_path):
     f = github_personal_input(tmp_path)
     cli(["integrations", "github", "add-personal", "-f", str(f)])
 
-def test_integrations_github_get():
+def test_integrations_github(tmp_path, capsys):
+    cli(["integrations", "github", "get", "-a", "github-test-3"])
+    out, err = capsys.readouterr()
+    out = json.loads(out)
+    if (out['type'] != "NOT_FOUND"):
+        cli(["integrations", "github", "delete", "-a", "github-test-3"])
+
+    f = github_app_input(tmp_path)
+    cli(["integrations", "github", "add", "-f", str(f)])
+
     cli(["integrations", "github", "get", "-a", "github-test-3"])
 
-def test_integrations_github_get_personal():
     cli(["integrations", "github", "get-personal", "-a", "github-personal-test-001"])
 
-def test_integrations_github_get_all():
     cli(["integrations", "github", "get-all"])
 
-def test_integrations_github_get_default():
     cli(["integrations", "github", "get-default"])
 
-def test_integrations_github_validate():
     cli(["integrations", "github", "validate", "-a", "cortex-test"])
 
-def test_integrations_github_validate_all():
     cli(["integrations", "github", "validate-all"])
 
-def test_integrations_github_update_personal(tmp_path):
     f = github_personal_input(tmp_path)
     cli(["integrations", "github", "update-personal", "-a", "github-personal-test-001", "-f", str(f)])
 
-def test_integrations_github_update():
     cli(["integrations", "github", "update", "-a", "github-test-3", "-f", "tests/test_integrations_github_update.json"])
 
 # Intentionally missing delete-all test because I didn't want to destroy my test env