+++ tests/test_resource_definitions.py
@@ -5,7 +5,7 @@ from cortexapps_cli.cortex import cli
 import json
 import sys
 
-def test_resource_definitions_create(capsys):
+def test_resource_definitions(capsys):
     # Delete resource definition if it already exists.
     cli(["resource-definitions", "list"])
     out, err = capsys.readouterr()
@@ -14,11 +14,8 @@ def test_resource_definitions_create(capsys):
         cli(["resource-definitions", "delete", "-t", "test-resource-definition"])
     cli(["resource-definitions", "create", "-f", "tests/test-resource-definition.json"])
 
-def test_resource_definitions_list():
     cli(["resource-definitions", "list"])
 
-def test_resource_definitions_get():
     cli(["resource-definitions", "get", "-t", "test-resource-definition"])
 
-def test_resource_definitions_update():
     cli(["resource-definitions", "update", "-t", "test-resource-definition", "-f", "tests/test-resource-definition-update.json"])